+++ packages/constants/CHANGELOG.md
@@ -215,6 +215,13 @@
   * devDependencies
     * @leather.io/prettier-config bumped to 0.9.0
 
+## [0.38.0](https://github.com/leather-io/mono/compare/@leather.io/constants-v0.37.2...@leather.io/constants-v0.38.0) (2026-08-05)
+
+
+### Features
+
+* add new activity list ([#2555](https://github.com/leather-io/mono/issues/2555)) ([7d5f88a](https://github.com/leather-io/mono/commit/7d5f88a8f5af565d381c231a401842462263c322))
+
 ## [0.37.2](https://github.com/leather-io/mono/compare/@leather.io/constants-v0.37.1...@leather.io/constants-v0.37.2) (2026-07-23)